widen the last three pointers squeezed into ints
Three sites left after the command payload and the Scintilla wrappers, each a
different mechanism.

SHTREEITEMHash hashed an HTREEITEM, which is a pointer, by casting it to int.
That is a worse hash on x64 rather than a wrong one, since the low half of a
heap address still spreads, but it is a truncation and std::hash<const void *>
already does the job. The functor returns size_t now, which is what
std::unordered_map wanted in the first place.

NMHDR::idFrom is a UINT_PTR, because it carries the control's HWND rather than
its id when TTF_IDISHWND is set. CResizeDialog::OnNeedToolTipText read it into
an unsigned and then handed the result back to GetDlgCtrlID as an HWND, which
on x64 is half a handle. It reads the UINT_PTR, resolves it to a control id,
and only then narrows, which is the point at which narrowing is safe.

Scintilla's ListBoxX::Create passes its control id in CreateWindowEx's HMENU
parameter, which is what Win32 asks for on a child window. A small int does
belong in that pointer slot; it just has to be widened to pointer size before
it can be spelled as a pointer.

With these the editor builds with no pointer truncation warning anywhere: 166
at the start of this pass, 0 now.

+++ Versions/Temporary/Engine/Sources/Scintilla/PlatWin.cxx
@@ -966,7 +966,7 @@ void ListBoxX::Create(Window &parent, int ctrlID, int lineHeight_, bool unicodeM
 		WS_EX_WINDOWEDGE, ListBoxX_ClassName, "",
 		WS_CHILD | WS_THICKFRAME,
 		100,100, 150,80, reinterpret_cast<HWND>(parent.GetID()),
-		reinterpret_cast<HMENU>(ctrlID),
+		reinterpret_cast<HMENU>(static_cast<uintptr_t>(ctrlID)),
 		hinstanceParent,
 		this);
 	lb = ::GetWindow(reinterpret_cast<HWND>(id), GW_CHILD);
